+++ src/simmate/apps/baderkit/models/__init__.py
@@ -2,14 +2,8 @@
 
 from .badelf import Badelf, BadelfCalculation
 from .bader import Bader
-from .elf_analysis import (
-    ElfAnalysis,
-    ElfAnalysisCalculation,
-)
+from .elf_analysis import ElfAnalysis, ElfAnalysisCalculation
 from .elf_features import ElfFeatures
 from .elf_radii import ElfRadii
 from .spin_badelf import SpinBadelf, SpinBadelfCalculation
-from .spin_elf_analysis import (
-    SpinElfAnalysis,
-    SpinElfAnalysisCalculation,
-)
+from .spin_elf_analysis import SpinElfAnalysis, SpinElfAnalysisCalculation