+++ src/simmate/apps/baderkit/models/elf_features.py
@@ -1,9 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from simmate.database.base_data_types import (
-    DatabaseTable,
-    table_column,
-)
+from simmate.database.base_data_types import DatabaseTable, table_column
 
 
 class ElfFeatures(DatabaseTable):