+++ src/simmate/apps/warren_lab/workflows/relaxation/hse.py
@@ -1,9 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from simmate.apps.warren_lab.inputs.potcar_mappings import HSE_POTCAR_MAPPINGS
-from simmate.apps.warren_lab.workflows.relaxation.pbe import (
-    Relaxation__Vasp__PbeWarren,
-)
+from simmate.apps.warren_lab.workflows.relaxation.pbe import Relaxation__Vasp__PbeWarren
 
 
 class Relaxation__Vasp__HseWarren(Relaxation__Vasp__PbeWarren):