+++ src/simmate/apps/warren_lab/workflows/staged/prepop_relaxation_static_hse_hse.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from simmate.apps.warren_lab.workflows.relaxation.hse import (
-    Relaxation__Vasp__HseWarren,
-)
+from simmate.apps.warren_lab.workflows.relaxation.hse import Relaxation__Vasp__HseWarren
 from simmate.workflows.base_flow_types import StagedWorkflow